// src/app/models/survey.ts
export class Survey {
    pid: String;
    interviewerId: String;
    firstName: String;
    lastName: String;
    age: String;
    gender: String;
    maritalStatus: String;
    educationLevel: String;
    country: String;
    province: String;
    district: String;
    tobaccoProducts: String;
    alcoholConsumption: String;
    eatFruit: String;
    highSalt: String;
    oilType: String;
}